requirements.txt
haystack-ai
mcp-haystack
load_dotenv
huggingface_hub
